# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a workout tracking application API",
        "title": "Workout Tracker API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/register": {
            "post": {
                "description": "Register new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Register new user",
                "parameters": [
                    {
                        "description": "User registration info",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/db.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/workout": {
            "post": {
                "description": "Create empty workout",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workout"
                ],
                "summary": "Create empty workout",
                "parameters": [
                    {
                        "description": "Worokout info",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/db.Workout"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "workout",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/workout/{workoutID}/exercise": {
            "post": {
                "description": "Add exercise to workout",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workout"
                ],
                "summary": "Add exercise to workout",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Workout ID",
                        "name": "workoutID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Exercise info",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/db.Exercise"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "exercise",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "db.Exercise": {
            "type": "object"
        },
        "db.User": {
            "type": "object"
        },
        "db.Workout": {
            "type": "object"
        }
    }
}